+++ speed/adolc/sparse_hessian.cpp
@@ -61,7 +61,7 @@ bool link_sparse_hessian(
    size_t capacity;           // capacity of an allocation
 
    // tape identifier
-   int tag  = 0;
+   short tag  = 0;
    // AD domain space vector
    ADVector a_x = thread_alloc::create_array<ADScalar>(n, capacity);
    // AD range space vector
@@ -85,10 +85,10 @@ bool link_sparse_hessian(
    // ----------------------------------------------------------------------
    if( ! global_option["onetape"] ) while(repeat--)
    {  // choose a value for x
-      CppAD::uniform_01(n, x);
+      CppAD::uniform_01( size_t(n), x);
 
       // declare independent variables
-      int keep = 0; // keep forward mode results
+      int   keep = 0; // keep forward mode results
       trace_on(tag, keep);
       for(size_t j = 0; j < n; j++)
          a_x[j] <<= x[j];
@@ -121,10 +121,10 @@ bool link_sparse_hessian(
    }
    else
    {  // choose a value for x
-      CppAD::uniform_01(n, x);
+      CppAD::uniform_01( size_t(n), x);
 
       // declare independent variables
-      int keep = 0; // keep forward mode results
+      int   keep = 0; // keep forward mode results
       trace_on(tag, keep);
       for(size_t j = 0; j < n; j++)
          a_x[j] <<= x[j];
@@ -141,7 +141,7 @@ bool link_sparse_hessian(
 
       while(repeat--)
       {  // choose a value for x
-         CppAD::uniform_01(n, x);
+         CppAD::uniform_01( size_t(n), x);
 
          // calculate the hessian at this x
          sparse_hess(tag, int(n),
